+++ src/components/DateRangePickerInputController.jsx
@@ -20,8 +20,6 @@ import toLocalizedDateString from '../utils/toLocalizedDateString';
 import isInclusivelyAfterDay from '../utils/isInclusivelyAfterDay';
 import isBeforeDay from '../utils/isBeforeDay';
 
-import BaseClass from '../utils/baseClass';
-
 import {
   START_DATE,
   END_DATE,
@@ -133,8 +131,7 @@ const defaultProps = {
   isRTL: false,
 };
 
-/** @extends React.Component */
-export default class DateRangePickerInputController extends BaseClass {
+export default class DateRangePickerInputController extends React.PureComponent {
   constructor(props) {
     super(props);